Show only relevant comments in the timeline
