+++ src/Module/Conversation/Channel.php
@@ -10,7 +10,8 @@ namespace Friendica\Module\Conversation;
 use Friendica\App;
 use Friendica\App\Mode;
 use Friendica\Content\BoundariesPager;
-use Friendica\Content\Conversation;
+use Friendica\Content\Conversation\ConversationRenderer;
+use Friendica\Content\Conversation\StatusEditor;
 use Friendica\Content\Conversation\Entity\Channel as ChannelEntity;
 use Friendica\Content\Conversation\Factory\UserDefinedChannel as UserDefinedChannelFactory;
 use Friendica\Content\Conversation\Factory\Timeline as TimelineFactory;
@@ -42,8 +43,10 @@ class Channel extends Timeline
 {
 	/** @var TimelineFactory */
 	protected $timeline;
-	/** @var Conversation */
-	protected $conversation;
+	/** @var ConversationRenderer */
+	protected $conversationRenderer;
+	/** @var StatusEditor */
+	protected $statusEditor;
 	/** @var App\Page */
 	protected $page;
 	/** @var SystemMessages */
@@ -57,18 +60,19 @@ class Channel extends Timeline
 	/** @var NetworkFactory */
 	protected $networkFactory;
 
-	public function __construct(UserDefinedChannelFactory $userDefinedChannel, NetworkFactory $network, CommunityFactory $community, ChannelFactory $channelFactory, ChannelRepository $channel, TimelineFactory $timeline, Conversation $conversation, App\Page $page, SystemMessages $systemMessages, Mode $mode, IHandleUserSessions $session, Database $database, IManagePersonalConfigValues $pConfig, IManageConfigValues $config, ICanCache $cache, ActivityFactory $ActivityFactory, L10n $l10n, App\BaseURL $baseUrl, App\Arguments $args, LoggerInterface $logger, Profiler $profiler, Response $response, array $server, array $parameters = [])
+	public function __construct(UserDefinedChannelFactory $userDefinedChannel, NetworkFactory $network, CommunityFactory $community, ChannelFactory $channelFactory, ChannelRepository $channel, TimelineFactory $timeline, ConversationRenderer $conversationRenderer, StatusEditor $statusEditor, App\Page $page, SystemMessages $systemMessages, Mode $mode, IHandleUserSessions $session, Database $database, IManagePersonalConfigValues $pConfig, IManageConfigValues $config, ICanCache $cache, ActivityFactory $ActivityFactory, L10n $l10n, App\BaseURL $baseUrl, App\Arguments $args, LoggerInterface $logger, Profiler $profiler, Response $response, array $server, array $parameters = [])
 	{
 		parent::__construct($channel, $mode, $session, $database, $pConfig, $config, $cache, $ActivityFactory, $l10n, $baseUrl, $args, $logger, $profiler, $response, $server, $parameters);
 
-		$this->timeline           = $timeline;
-		$this->conversation       = $conversation;
-		$this->page               = $page;
-		$this->systemMessages     = $systemMessages;
-		$this->channel            = $channelFactory;
-		$this->community          = $community;
-		$this->networkFactory     = $network;
-		$this->userDefinedChannel = $userDefinedChannel;
+		$this->timeline             = $timeline;
+		$this->conversationRenderer = $conversationRenderer;
+		$this->statusEditor         = $statusEditor;
+		$this->page                 = $page;
+		$this->systemMessages       = $systemMessages;
+		$this->channel              = $channelFactory;
+		$this->community            = $community;
+		$this->networkFactory       = $network;
+		$this->userDefinedChannel   = $userDefinedChannel;
 	}
 
 	protected function content(array $request = []): string
@@ -106,7 +110,7 @@ class Channel extends Timeline
 			}
 
 			// We need the editor here to be able to reshare an item.
-			$o .= $this->conversation->statusEditor([], 0, true);
+			$o .= $this->statusEditor->renderEditor([], 0, true);
 		}
 
 		if ($this->channel->isTimeline($this->selectedTab) || $this->userDefinedChannel->isTimeline($this->selectedTab, $this->session->getLocalUserId())) {
@@ -122,7 +126,7 @@ class Channel extends Timeline
 			return $o;
 		}
 
-		$o .= $this->conversation->render($items, Conversation::MODE_CHANNEL, $this->raw, false, $order, $this->session->getLocalUserId());
+		$o .= $this->conversationRenderer->renderThreaded($items, ConversationRenderer::MODE_CHANNEL, $this->raw, $order, $this->session->getLocalUserId());
 
 		$pager = new BoundariesPager(
 			$this->l10n,
