+++ src/Module/Conversation/Community.php
@@ -13,7 +13,8 @@ namespace Friendica\Module\Conversation;
 use Friendica\App;
 use Friendica\App\Mode;
 use Friendica\Content\BoundariesPager;
-use Friendica\Content\Conversation;
+use Friendica\Content\Conversation\ConversationRenderer;
+use Friendica\Content\Conversation\StatusEditor;
 use Friendica\Content\Conversation\Entity\Community as CommunityEntity;
 use Friendica\Content\Conversation\Factory\Community as CommunityFactory;
 use Friendica\Content\Conversation\Factory\Activity as ActivityFactory;
@@ -52,21 +53,24 @@ class Community extends Timeline
 
 	/** @var CommunityFactory */
 	protected $community;
-	/** @var Conversation */
-	protected $conversation;
+	/** @var ConversationRenderer */
+	protected $conversationRenderer;
+	/** @var StatusEditor */
+	protected $statusEditor;
 	/** @var App\Page */
 	protected $page;
 	/** @var SystemMessages */
 	protected $systemMessages;
 
-	public function __construct(UserDefinedChannel $channel, CommunityFactory $community, Conversation $conversation, App\Page $page, SystemMessages $systemMessages, Mode $mode, IHandleUserSessions $session, Database $database, IManagePersonalConfigValues $pConfig, IManageConfigValues $config, ICanCache $cache, ActivityFactory $ActivityFactory, L10n $l10n, App\BaseURL $baseUrl, App\Arguments $args, LoggerInterface $logger, Profiler $profiler, Response $response, array $server, array $parameters = [])
+	public function __construct(UserDefinedChannel $channel, CommunityFactory $community, ConversationRenderer $conversationRenderer, StatusEditor $statusEditor, App\Page $page, SystemMessages $systemMessages, Mode $mode, IHandleUserSessions $session, Database $database, IManagePersonalConfigValues $pConfig, IManageConfigValues $config, ICanCache $cache, ActivityFactory $ActivityFactory, L10n $l10n, App\BaseURL $baseUrl, App\Arguments $args, LoggerInterface $logger, Profiler $profiler, Response $response, array $server, array $parameters = [])
 	{
 		parent::__construct($channel, $mode, $session, $database, $pConfig, $config, $cache, $ActivityFactory, $l10n, $baseUrl, $args, $logger, $profiler, $response, $server, $parameters);
 
-		$this->community      = $community;
-		$this->conversation   = $conversation;
-		$this->page           = $page;
-		$this->systemMessages = $systemMessages;
+		$this->community            = $community;
+		$this->conversationRenderer = $conversationRenderer;
+		$this->statusEditor         = $statusEditor;
+		$this->page                 = $page;
+		$this->systemMessages       = $systemMessages;
 	}
 
 	protected function content(array $request = []): string
@@ -100,7 +104,7 @@ class Community extends Timeline
 
 			// We need the editor here to be able to reshare an item.
 			if ($this->session->isAuthenticated()) {
-				$o .= $this->conversation->statusEditor([], 0, true);
+				$o .= $this->statusEditor->renderEditor([], 0, true);
 			}
 		}
 
@@ -113,7 +117,7 @@ class Community extends Timeline
 			return $o;
 		}
 
-		$o .= $this->conversation->render($items, Conversation::MODE_COMMUNITY, $this->raw, false, 'received', $this->session->getLocalUserId());
+		$o .= $this->conversationRenderer->renderThreaded($items, ConversationRenderer::MODE_COMMUNITY, $this->raw, ConversationRenderer::ORDER_RECEIVED, $this->session->getLocalUserId());
 
 		$pager = new BoundariesPager(
 			$this->l10n,
