+++ src/Module/Update/Notes.php
@@ -10,7 +10,7 @@
 
 namespace Friendica\Module\Update;
 
-use Friendica\Content\Conversation;
+use Friendica\Content\Conversation\ConversationRenderer;
 use Friendica\Core\System;
 use Friendica\Database\DBA;
 use Friendica\Model\Item;
@@ -57,6 +57,6 @@ final class Notes extends ProfileModule
 			throw new NotFoundException();
 		}
 
-		System::htmlUpdateExit($this->conversation->render(Post::toArray($r), Conversation::MODE_NOTES, true));
+		System::htmlUpdateExit($this->conversationRenderer->renderThreaded(Post::toArray($r), ConversationRenderer::MODE_NOTES, true));
 	}
 }
