+++ static/dbview.config.php
@@ -53,6 +53,7 @@ return [
 			"uid"          => ["channel-post", "uid"],
 			"uri-id"       => ["channel-post", "uri-id"],
 			"in-timeline"  => ["channel-post", "in-timeline"],
+			"author-id"    => ["post-thread", "author-id"],
 			"contact-id"   => ["post-engagement", "owner-id"],
 			"owner-id"     => ["post-engagement", "owner-id"],
 			"media-type"   => ["post-engagement", "media-type"],
@@ -82,6 +83,7 @@ return [
 			"uid"          => ["system-channel-post", "uid"],
 			"uri-id"       => ["system-channel-post", "uri-id"],
 			"in-timeline"  => ["system-channel-post", "in-timeline"],
+			"author-id"    => ["post-thread", "author-id"],
 			"contact-id"   => ["post-engagement", "owner-id"],
 			"owner-id"     => ["post-engagement", "owner-id"],
 			"media-type"   => ["post-engagement", "media-type"],
@@ -1505,6 +1507,7 @@ return [
 			"mention"      => ["post-thread-user", "mention"],
 			"network"      => ["post-thread-user", "network"],
 			"protocol"     => ["post-user", "protocol"],
+			"author-id"    => ["post-thread-user", "author-id"],
 			"contact-id"   => ["post-thread-user", "contact-id"],
 			"contact-type" => ["ownercontact", "contact-type"],
 		],
